Collect week-calendar utility functions in src/lib/units/week-calendar-utils
