+++ src/lib/units/week.js
@@ -4,6 +4,7 @@ import { addRegexToken, match1to2, match2 } from '../parse/regex';
 import { addWeekParseToken } from '../parse/token';
 import toInt from '../utils/to-int';
 import { createLocal } from '../create/local';
+import { weekOfYear } from './week-calendar-utils';
 
 // FORMATTING
 
@@ -27,34 +28,6 @@ addWeekParseToken(['w', 'ww', 'W', 'WW'], function (input, week, config, token) 
 });
 
 // HELPERS
-
-// firstDayOfWeek       0 = sun, 6 = sat
-//                      the day of the week that starts the week
-//                      (usually sunday or monday)
-// firstDayOfWeekOfYear 0 = sun, 6 = sat
-//                      the first week is the week that contains the first
-//                      of this day of the week
-//                      (eg. ISO weeks use thursday (4))
-export function weekOfYear(mom, firstDayOfWeek, firstDayOfWeekOfYear) {
-    var end = firstDayOfWeekOfYear - firstDayOfWeek,
-        daysToDayOfWeek = firstDayOfWeekOfYear - mom.day(),
-        adjustedMoment;
-
-
-    if (daysToDayOfWeek > end) {
-        daysToDayOfWeek -= 7;
-    }
-
-    if (daysToDayOfWeek < end - 7) {
-        daysToDayOfWeek += 7;
-    }
-
-    adjustedMoment = createLocal(mom).add(daysToDayOfWeek, 'd');
-    return {
-        week: Math.ceil(adjustedMoment.dayOfYear() / 7),
-        year: adjustedMoment.year()
-    };
-}
 
 // LOCALES
